Chore: Move usePlugins closer to roles pages
