+++ packages/plugins/users-permissions/admin/src/pages/Roles/pages/CreatePage.js
@@ -27,9 +27,9 @@ import { useMutation } from 'react-query';
 import { useHistory } from 'react-router-dom';
 
 import UsersPermissions from '../../../components/UsersPermissions';
-import { usePlugins } from '../../../hooks/usePlugins';
 import getTrad from '../../../utils/getTrad';
 import { createRoleSchema } from '../constants';
+import { usePlugins } from '../hooks/usePlugins';
 
 export const CreatePage = () => {
   const { formatMessage } = useIntl();
